@@ -1,6 +1,6 @@
 package com.example.new_sample;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.junit.Assert.*;
 
